feat: mutate union types
